+++ opaca-model/src/main/java/de/gtarc/opaca/model/AgentContainerImage.java
@@ -45,6 +45,9 @@ public class AgentContainerImage {
     /** provider of the container, e.g. institute or researcher */
     String provider;
 
+    /** URL to this image's repository (public, if any, or private), website, or other source for more information */
+    String url;
+
     // OPTIONAL attributes for API port (if not default) and extra ports (if any)
 
     /** the port where the container provides the OPACA API; by default this is 8082 but another may be used */